Add files via upload
This is Basic Movement file. The logic controls movement of two different motors. The servo is used for steering mechanism, so that robot will move left and right.
Camera Functionality: For this code we are not using
Sonar sensor: For this code we are not using
The Code is written together by Jaysimmha, Sai and Vishwa